% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppi.R
\name{calculate_ppi}
\alias{calculate_ppi}
\title{Calculate PPI for a set of sites or grid cells given one or more buffer distances}
\usage{
calculate_ppi(locs, buffer, compiled_raster, progress = FALSE)
}
\arguments{
\item{locs}{An \code{sf} object of geospatial features (typically POINTS or POLYGONS).}

\item{buffer}{A numeric vector of one or more buffer distances (in kilometers) to use in focal mean calculations.}

\item{compiled_raster}{A \code{terra} raster object containing population densities.}

\item{progress}{Logical; if TRUE, shows progress bar during \code{exactextractr::exact_extract()} operations.}
}
\value{
An \code{sf} object with additional columns for each PPI buffer (e.g., PPI-10km).
}
\description{
This function computes the focal mean of population density within one or more buffer distances
around a set of input locations or polygons, using the exactextractr package for accurate spatial summaries.
}
