% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rasters.R
\name{compile_rasters}
\alias{compile_rasters}
\title{Import raster tiles for the study area of interest and (if more than one) assemble into a single, mosaiced spatial data layer.}
\usage{
compile_rasters(
  tileIDs,
  study_area,
  data_directory = "data",
  study_area_buffer = 1000
)
}
\arguments{
\item{tileIDs}{vector of tile_ids for selected GHS raster tiles}

\item{study_area}{sf object of the study area grid in Molleweide projection to match the coordinate reference system of raster tiles}

\item{data_directory}{Directory where GHS raster files have been downloaded and unzipped}

\item{study_area_buffer}{Extra buffer around the study area to which population density raster should be cropped. Default = 1km.}
}
\value{
A raster layer of population densities.
}
\description{
Import raster tiles for the study area of interest and (if more than one) assemble into a single, mosaiced spatial data layer.
}
