% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tile_mosaics.R
\name{create_tile_mosaics}
\alias{create_tile_mosaics}
\title{Create raster mosaics for unique tile combinations}
\usage{
create_tile_mosaics(
  tile_list,
  data_directory,
  tif_prefix = "GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100_V1_0_"
)
}
\arguments{
\item{tile_list}{A list of character vectors (as output by get_tile_list)}

\item{data_directory}{Directory where "downloads" and "mosaics" folders are/will be created}

\item{tif_prefix}{Prefix used in .tif file naming}
}
\value{
A character vector of created mosaic file paths (empty if none were created)
}
\description{
Create raster mosaics for unique tile combinations
}
