% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghs_tiles.R
\name{download_ghs_tiles}
\alias{download_ghs_tiles}
\title{Check for missing GHS raster tiles and download them if needed}
\usage{
download_ghs_tiles(
  tile_list,
  data_directory,
  URL = NULL,
  ghs_layer = "BUILT_S",
  ghs_year = 2030,
  ghs_epsg = 54009,
  ghs_resolution = "100"
)
}
\arguments{
\item{tile_list}{A list of tile ID vectors (e.g., output from get_tile_list)}

\item{data_directory}{Directory where "downloads" folder will be created and tiles stored}

\item{URL}{Base URL of the GHSL raster zip file library}

\item{ghs_layer}{GHS layer of interest (default: BUILT_S)}

\item{ghs_year}{GHS layer year of interest (default: 2030)}

\item{ghs_epsg}{GHS layer coordinate reference system of interest (default: 54009 - Mollweide)}

\item{ghs_resolution}{GHS layer resolution of interest (default: 100 meters)}
}
\value{
A named logical vector indicating whether each tile is now present
}
\description{
Check for missing GHS raster tiles and download them if needed
}
