% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Check for missing GHS raster tiles and download them if needed}
\usage{
download_tiles(
  tile_list,
  data_directory,
  URL = NULL,
  tif_prefix = "GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100_V1_0_"
)
}
\arguments{
\item{tile_list}{A list of tile ID vectors (e.g., output from get_tile_list)}

\item{data_directory}{Directory where "downloads" folder will be created and tiles stored}

\item{URL}{Base URL of the GHSL raster zip file library}

\item{tif_prefix}{Prefix used in naming the .tif files}
}
\value{
A named logical vector indicating whether each tile is now present
}
\description{
Check for missing GHS raster tiles and download them if needed
}
