% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{EasternAustraliaSites}
\alias{EasternAustraliaSites}
\title{Made-Up Marine Sites in Eastern Australia}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{Longitude}{Longitude in decimal degrees (E)}
\item{Latitude}{Latitude in decimal degrees (S)}
}
}
\usage{
EasternAustraliaSites
}
\description{
A dataset of arbitrary geospatial coordinates along the Eastern Australian coastline.
}
\keyword{datasets}
