% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_rasters}
\alias{get_rasters}
\title{Load the correct raster (single or mosaic) for each location from tile_list, with memoised loading}
\usage{
get_rasters(
  tile_list,
  data_directory,
  tif_prefix = "GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100_V1_0_",
  return_paths = FALSE
)
}
\arguments{
\item{tile_list}{A list of tile ID vectors (from get_tile_list)}

\item{data_directory}{The base directory containing "downloads/" and "mosaics/" folders}

\item{tif_prefix}{Prefix used in naming the .tif files}

\item{return_paths}{Logical. If TRUE, returns file paths instead of loading rasters. Default is FALSE}
}
\value{
A list of either terra::rast objects or character file paths, one per entry in tile_list
}
\description{
Load the correct raster (single or mosaic) for each location from tile_list, with memoised loading
}
