% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_schema.R
\name{get_tile_schema}
\alias{get_tile_schema}
\title{Download and load the GHSL tile schema shapefile (land tiles only)}
\usage{
get_tile_schema(
  data_directory,
  tile_schema_URL =
    "https://ghsl.jrc.ec.europa.eu/download/GHSL_data_54009_shapefile.zip"
)
}
\arguments{
\item{data_directory}{A character string specifying the directory where the tile schema
shapefile should be stored. If it is not present, it will be downloaded here.}

\item{tile_schema_URL}{A character string specifying the URL of the zipped tile schema
shapefile. Defaults to the GHSL land tile schema URL.}
}
\value{
An \code{sf} object containing the tile schema polygons with a \code{tile_id} column.
}
\description{
This function checks for the GHSL tile schema shapefile in a given local directory.
If the shapefile is missing, it downloads and unzips the schema from the official
GHSL (Global Human Settlement Layer) repository. It returns the tile schema as an \code{sf} object.
}
\details{
The function ensures that the returned object includes a \code{tile_id} column, which is
required for matching spatial buffers to the correct raster tiles. It also filters to
land tiles only, based on filenames that include \code{"tile_schema_land.shp"}.
}
\examples{
\dontrun{
schema <- get_tile_schema(data_directory = "Reference Files/rasters")
plot(schema["tile_id"])
}

}
