% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_to_dd.R
\name{km_to_dd}
\alias{km_to_dd}
\title{Convert km to longitude and latitude decimal degrees based on great circle distance}
\usage{
km_to_dd(km, lat_dd)
}
\arguments{
\item{km}{A distance in kilometers that will be converted to decimal degrees in 2 dimensions.}

\item{lat_dd}{A vector of latitudes in decimal degrees for the study area.}
}
\value{
A data frame of longitudinal and latitudinal equivalencies to input distance.
}
\description{
Convert km to longitude and latitude decimal degrees based on great circle distance
}
\examples{
km_to_dd(2, 48.84)
}
