% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Create a grid of empty x by x km cells for which to calculate PPI.}
\usage{
make_grid(locs, buffer, dimensions = 0.5, bounds = NULL)
}
\arguments{
\item{locs}{An sf object of grid cells}

\item{buffer}{A vector of buffer distances that should be used to compute focal means for population density.}

\item{dimensions}{A number specifying the desired dimensions of square grid cells in kilometers}

\item{bounds}{If preferred, the grid can be calculated with specific bounds by providing a vector of bounding values here in the format, c(xmin, ymin, xmax, ymax)}
}
\value{
An sf object of grid cell polygons.
}
\description{
Create a grid of empty x by x km cells for which to calculate PPI.
}
\section{Details}{

Are they ungulates?
}

