% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarinePPI.R
\name{MarinePPI}
\alias{MarinePPI}
\title{Wrapper function for calculating PPI}
\usage{
MarinePPI(
  locs,
  buffer,
  getgridcalcs = FALSE,
  dimensions = 0.5,
  bounds = NULL,
  data_directory = "data",
  tile_schema_URL =
    "https://ghsl.jrc.ec.europa.eu/download/GHSL_data_54009_shapefile.zip",
  tile_library_URL =
    "https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_BUILT_S_GLOBE_R2023A/GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100/V1-0/tiles/",
  study_area_buffer = 1000,
  progress = FALSE
)
}
\arguments{
\item{locs}{An sf object of geospatial coordinates (POINTS) or grid cells (POLYGONS).}

\item{buffer}{A vector of one or more buffer distance(s) that should be in focal mean calculations, expressed in kilometers.}

\item{getgridcalcs}{logical; do you want to return PPI estimates for the entire grid?}

\item{dimensions}{A number specifying the desired dimensions of square grid cells in kilometers; default is 0.5 x 0.5 km.}

\item{bounds}{If preferred, the grid can be calculated with specific bounds by providing a vector of bounding values here in the format, c(xmin, ymin, xmax, ymax)}

\item{data_directory}{Data directory where GHS files should be stored.}

\item{tile_schema_URL}{url for the shapefile of GHS tile areas}

\item{tile_library_URL}{url for the GHS library containing individual zipfiles of each raster tile}

\item{study_area_buffer}{Extra buffer around the study area to which population density raster should be cropped. Default = 1km.}

\item{progress}{if TRUE, display a progress bar while running exactextractr::exact_extract()}
}
\value{
sf objects containing calculated values of PPI for each sites and/or grid cell (depending on the how the argument outputtype is specified).
}
\description{
Wrapper function for calculating PPI
}
