% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PugetSoundSites.R
\docType{data}
\name{PugetSoundSites}
\alias{PugetSoundSites}
\title{Puget Sound Sites}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{Location}{Site name}
\item{Lat_dd}{Latitude in decimal degrees}
\item{Lat_dd}{Longitude in decimal degrees}
}on
}
\usage{
PugetSoundSites
}
\description{
Sample data frame of geospatial coordinates for 22 sites in Puget Sound
}
\keyword{datasets}
