% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppi.R
\name{calculate_ppi}
\alias{calculate_ppi}
\title{Calculate PPI a set of sites or grid cells given one or more buffer distances.}
\usage{
calculate_ppi(locs, buffer, compiled_raster, progress = FALSE)
}
\arguments{
\item{locs}{An sf object of geospatial coordinates (POINTS) or grid cells (POLYGONS).}

\item{buffer}{A vector of one or more buffer distance(s) that should be in focal mean calculations, expressed in kilometers.}

\item{compiled_raster}{A raster of population densities}

\item{progress}{if TRUE, display a progress bar while running exactextractr::exact_extract()}
}
\value{
An sf object of PPI estimates for site.
}
\description{
Calculate PPI a set of sites or grid cells given one or more buffer distances.
}
