% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rasters.R
\name{compile_rasters}
\alias{compile_rasters}
\title{Import raster tiles for the study area of interest and (if more than one) assemble into a single, mosaiced spatial data layer.}
\usage{
compile_rasters(tiles, data_directory = "downloads", study_area_buffer = 1000)
}
\arguments{
\item{tiles}{output from the get_ghs_tiles() function}

\item{data_directory}{Directory where GHS raster files have been downloaded and unzipped}

\item{study_area_buffer}{Extra buffer around the study area to which population density raster should be cropped. Default = 1km.}
}
\value{
A raster layer of population densities.
}
\description{
Import raster tiles for the study area of interest and (if more than one) assemble into a single, mosaiced spatial data layer.
}
