% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghs_tiles.R
\name{download_ghs_tiles}
\alias{download_ghs_tiles}
\title{Download and unzip raster tiles for the study area of interest.}
\usage{
download_ghs_tiles(
  tileIDs,
  data_directory = getwd(),
  URL =
    "https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_BUILT_S_GLOBE_R2023A/GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100/V1-0/tiles/"
)
}
\arguments{
\item{tileIDs}{vector of tile_ids for selected GHS raster tiles}

\item{data_directory}{Directory where GHS raster files have been downloaded and unzipped}

\item{URL}{Weblink for the GHS library of tiled raster zip files}
}
\value{
NA
}
\description{
Download and unzip raster tiles for the study area of interest.
}
