% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghs_tiles.R
\name{get_ghs_tiles}
\alias{get_ghs_tiles}
\title{Identify and access GHS tiles necessary for assembling a single population density raster for the specified study area}
\usage{
get_ghs_tiles(
  study_area,
  data_directory = "downloads",
  tile_schema_URL =
    "https://ghsl.jrc.ec.europa.eu/download/GHSL_data_54009_shapefile.zip",
  tile_library_URL =
    "https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_BUILT_S_GLOBE_R2023A/GHS_BUILT_S_E2030_GLOBE_R2023A_54009_100/V1-0/tiles/"
)
}
\arguments{
\item{study_area}{Grid generated for study area}

\item{data_directory}{Directory where GHS raster files have been downloaded and unzipped}

\item{tile_schema_URL}{url for the shapefile of GHS tile areas}

\item{tile_library_URL}{url for the GHS library containing individual zipfiles of each raster tile}
}
\value{
list of three objects: (1) a data frame showing the tiles that were downloaded, (2) an sf POLYGON object of tile areas, and (3) an sf object of the study area grid, converted to Mollweide (the coordinate reference system of the imported rasters)
}
\description{
Identify and access GHS tiles necessary for assembling a single population density raster for the specified study area
}
