% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{PugetSoundSites}
\alias{PugetSoundSites}
\title{Puget Sound Sampling Locations}
\format{
A data frame with N rows and 3 variables:
\describe{
\item{Lon_dd}{Longitude in decimal degrees}
\item{Lat_dd}{Latitude in decimal degrees}
\item{Location}{Site name}
}
}
\usage{
PugetSoundSites
}
\description{
A dataset of site coordinates and names for 22 marine field sites in the Salish Sea.
}
\keyword{datasets}
